# api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "API productos de k-pop",
    "description": "API para gestionar álbumes de k-pop y usuarios en productosNova",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/swagger.json": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/health": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/albums/stats": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/albums/search": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          },
          {
            "name": "precioMin",
            "in": "query",
            "type": "string"
          },
          {
            "name": "precioMax",
            "in": "query",
            "type": "string"
          },
          {
            "name": "q",
            "in": "query",
            "type": "string"
          },
          {
            "name": "categoria",
            "in": "query",
            "type": "string"
          },
          {
            "name": "artista",
            "in": "query",
            "type": "string"
          },
          {
            "name": "disponible",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/albums/artista/{artista}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "artista",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/albums/categoria/{categoria}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "categoria",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/albums/": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "type": "string"
          },
          {
            "name": "limit",
            "in": "query",
            "type": "string"
          },
          {
            "name": "precioMin",
            "in": "query",
            "type": "string"
          },
          {
            "name": "precioMax",
            "in": "query",
            "type": "string"
          },
          {
            "name": "artista",
            "in": "query",
            "type": "string"
          },
          {
            "name": "categoria",
            "in": "query",
            "type": "string"
          },
          {
            "name": "disponible",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "fotoAlbum": {
                  "example": "any"
                },
                "artistaGrupo": {
                  "example": "any"
                },
                "artista": {
                  "example": "any"
                },
                "version": {
                  "example": "any"
                },
                "versionAlbum": {
                  "example": "any"
                },
                "peso": {
                  "example": "any"
                },
                "pesoGramos": {
                  "example": "any"
                },
                "fechaCompra": {
                  "example": "any"
                },
                "fechaAdquisicion": {
                  "example": "any"
                },
                "fechaCaducidad": {
                  "example": "any"
                },
                "fechaLimiteVenta": {
                  "example": "any"
                },
                "idioma": {
                  "example": "any"
                },
                "categoria": {
                  "example": "any"
                },
                "precio": {
                  "example": "any"
                },
                "nombreAlbum": {
                  "example": "any"
                },
                "duracion": {
                  "example": "any"
                },
                "stock": {
                  "example": "any"
                },
                "descripcion": {
                  "example": "any"
                },
                "fechaLanzamiento": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/albums/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "fotoAlbum": {
                  "example": "any"
                },
                "artistaGrupo": {
                  "example": "any"
                },
                "artista": {
                  "example": "any"
                },
                "version": {
                  "example": "any"
                },
                "versionAlbum": {
                  "example": "any"
                },
                "peso": {
                  "example": "any"
                },
                "pesoGramos": {
                  "example": "any"
                },
                "fechaCompra": {
                  "example": "any"
                },
                "fechaAdquisicion": {
                  "example": "any"
                },
                "fechaCaducidad": {
                  "example": "any"
                },
                "fechaLimiteVenta": {
                  "example": "any"
                },
                "idioma": {
                  "example": "any"
                },
                "categoria": {
                  "example": "any"
                },
                "precio": {
                  "example": "any"
                },
                "nombreAlbum": {
                  "example": "any"
                },
                "duracion": {
                  "example": "any"
                },
                "stock": {
                  "example": "any"
                },
                "descripcion": {
                  "example": "any"
                },
                "fechaLanzamiento": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/users/": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      },
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "contrasena": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                },
                "nombre": {
                  "example": "any"
                },
                "apellido": {
                  "example": "any"
                },
                "edad": {
                  "example": "any"
                },
                "numeroTelefono": {
                  "example": "any"
                },
                "rol": {
                  "example": "any"
                },
                "correo": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "patch": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "description": "",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "correo": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/auth/logout": {
      "post": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/auth/me": {
      "get": {
        "description": "",
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  }
}